backend/requirements.txt
fastapi
uvicorn
pytesseract
pdfplumber
openai
python-multipart
pillow
python-dotenv